--- backend/models/beehives.go
@@ -1,6 +1,0 @@
-package models
-
-type Beehives struct {
-	Id   int    `json: "id"`
-	Name string `json: "name`
-}
