--- backend/models/user.go
@@ -1,7 +1,0 @@
-package models
-
-// User is used to handle login.
-type User struct {
-	Username string `json:"username"`
-	Password string `json:"password"`
-}
